Upgraded the react google maps package   - now a supported module
