+++ package.json
@@ -26,6 +26,7 @@
     "@fortawesome/fontawesome": "^1.1.8",
     "@fortawesome/fontawesome-free-solid": "^5.0.13",
     "@fortawesome/react-fontawesome": "0.0.18",
+    "@react-google-maps/api": "^1.9.12",
     "animated-scroll-to": "^1.3.1",
     "axios": "^0.18.1",
     "babel-jest": "^24.9.0",
@@ -33,7 +34,6 @@
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-google-maps": "^9.4.5",
     "react-places-autocomplete": "^7.3.0",
     "react-redux": "^5.1.2",
     "redux": "^4.0.5",
